+++ Tests/NukeThreadSafetyTests/ThreadSafetyTests.swift
@@ -219,7 +219,6 @@ struct RandomizedTests {
         let pipeline = ImagePipeline {
             $0.dataLoader = dataLoader
             $0.imageCache = nil
-            $0.isRateLimiterEnabled = false
         }
 
         let queue = OperationQueue()